% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jum_c.R
\name{jum_c}
\alias{jum_c}
\title{Simple Morphological Analysis}
\usage{
jum_c(input, pos = NULL, redirect = FALSE)
}
\arguments{
\item{input}{input text}

\item{pos}{extract pattern i.e. "名詞|動詞"}

\item{redirect}{Whether or not redirect to Wikipedia redirect. Default is FALSE.}
}
\value{
named list of Morphological Analysis
}
\description{
Simple Morphological Analysis
}
