% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{jum_close_server}
\alias{jum_close_server}
\title{Close JUMAN++ server.}
\usage{
jum_close_server(handle)
}
\arguments{
\item{handle}{Process handle obtained from \code{jum_start_server()}}
}
\description{
Close JUMAN++ server.
}
