% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{jum_file}
\alias{jum_file}
\title{Simple result of JUMAN++(file)}
\usage{
jum_file(filename)
}
\arguments{
\item{filename}{input file}
}
\value{
result of JUMAN++
}
\description{
Simple result of JUMAN++(file)
}
