% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{jum_start_server}
\alias{jum_start_server}
\title{Start JUMAN++ server.}
\usage{
jum_start_server(host.name = NULL, port = NULL)
}
\arguments{
\item{host.name}{host name if necessary}

\item{port}{port number if necessary. Default is 12000}
}
\description{
Start JUMAN++ server.
}
