% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjumanpp_c.R
\name{rjumanpp_c}
\alias{rjumanpp_c}
\title{Simple Morphological Analysis}
\usage{
rjumanpp_c(input)
}
\arguments{
\item{input}{input text}
}
\value{
named list of Morphological Analysis
}
\description{
Simple Morphological Analysis
}
