% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjumnapp_wakati.R
\name{rjumanpp_wakati}
\alias{rjumanpp_wakati}
\title{jumanpp wakatigaki}
\usage{
rjumanpp_wakati(input, pattern = NULL)
}
\arguments{
\item{input}{input text}

\item{pattern}{extract pattern i.e. "名詞|動詞"}
}
\value{
seperated words of the input text
}
\description{
jumanpp wakatigaki
}
