% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jum_c.R
\name{jum_c}
\alias{jum_c}
\title{Simple Morphological Analysis}
\usage{
jum_c(input, mypref = 0, pos = NULL, redirect = FALSE, server = FALSE)
}
\arguments{
\item{input}{input text}

\item{mypref}{Default being 0, the same morphemic forms that appear on the text are returned. If 1 is designated, the basic forms of them are instead.}

\item{pos}{extract pattern i.e. "名詞|動詞"}

\item{redirect}{Whether or not redirect to Wikipedia redirect. Default is FALSE.}

\item{server}{if TRUE, JUMAN++ server is used. In such a case, you have to \command{jum_start_server} to start JUMAN++ server.}
}
\value{
named list of Morphological Analysis
}
\description{
Simple Morphological Analysis
}
