% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jum_c.R
\name{jum_c}
\alias{jum_c}
\title{Simple Morphological Analysis}
\usage{
jum_c(input)
}
\arguments{
\item{input}{input text}
}
\value{
named list of Morphological Analysis
}
\description{
Simple Morphological Analysis
}
