% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{jum_file}
\alias{jum_file}
\title{Simple result of JUMAN++(file)}
\usage{
jum_file(filename, server = FALSE)
}
\arguments{
\item{filename}{input file}

\item{server}{if TRUE, JUMAN++ server is used. In such a case, you have to \command{jum_start_server} to start JUMAN++ server.}
}
\value{
result of JUMAN++
}
\description{
Simple result of JUMAN++(file)
}
