% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jum_wakati.R
\name{jum_wakati}
\alias{jum_wakati}
\title{jumanpp wakatigaki}
\usage{
jum_wakati(input, pattern = NULL)
}
\arguments{
\item{input}{input text}

\item{pattern}{extract pattern i.e. "名詞|動詞"}
}
\value{
seperated words of the input text
}
\description{
jumanpp wakatigaki
}
